+++ custom_addons/object_request/__manifest__.py
@@ -23,6 +23,7 @@
         "stock",
         "purchase",
         "contacts",
+        "custom_product_search",
     ],
     "data": [
         "data/ir_sequence_data.xml",